+++ core/src/sync/message/new_block.rs
@@ -44,7 +44,7 @@ impl Handleable for NewBlock {
 
         ctx.manager.request_block_headers(
             ctx.io,
-            Some(ctx.peer),
+            Some(ctx.node_id.clone()),
             headers_to_request,
             true, /* ignore_db */
         );