+++ tests/test_framework/test_framework.py
@@ -546,9 +546,12 @@ class OptionHelper:
     def parsed_options_to_args(parsed_arg: dict):
         args = []
         for arg_name, value in parsed_arg.items():
-            args.append(OptionHelper.to_argument_str(arg_name))
             if type(value) is not bool:
+                args.append(OptionHelper.to_argument_str(arg_name))
                 args.append(str(value))
+            elif value:
+                # FIXME: This only allows setting boolean to True.
+                args.append(OptionHelper.to_argument_str(arg_name))
         return args
 
     """